requirements.txt
langchain>=0.3.13
langchain-openai>=0.2.14
langchain_community>=0.3.13
langgraph>=0.2.60
python-dotenv
ipykernel
streamlit
beautifulsoup4